pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "akro-agent"
version = "0.1.0"
description = "Multi-agent deep research system"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "openai>=1.12.0",
    "tavily-python>=0.5.0",
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.26.0",
    "reportlab>=4.0.0",
    "python-pptx>=0.6.21",
]

[project.optional-dependencies]
dev = []

[tool.setuptools.packages.find]
where = ["src"]
